split generate_inputs for T1w and dwi
Previously, if dwi_dir was passed, a naming scheme as assumed. Rather
than hardcode the naming scheme, this change passes dwi_dir through to
generate_inputs, which should enable different name schemes to be picked
up. Bonus, it suppresses the message of not being able to find files,
unless it actually is unable to find data in the provided paths. By
default, if dwi_dir is not provided, it falls back on bids_dir.