update config wildcards and filtering
Previously, the usage of `wildcards` was vast in the config file,
allowing for wildcards for all sorts of entities. This PR constrains the
wildcards to just subject, session, and run (can consider others in the
future, but these make the most sense at the moment).

Dwi filters are constrained to be in the "T1w" space, as we assume dwi
processing should be performed in the same space as T1w. Alternatively,
can consider "orig" space as well, but will need to think about to
handle transformations to the T1w in that case.

If multiple entities are found (e.g. "res"), Snakebids will throw an
error reflecting the files found and suggest to the user to make use of
the `--filter` option to constrain the files.